daq_logger refactoring using CustomApp base class
The effective used logger is now factory set from external objects

These objects must implement PyMoDAQ's logger interface enforced from inheritance of the Abstract Class : AbstractLogger

This means other logger objects implementing this interface can be added to the logger factory

+++ src/pymodaq/daq_scan.py
@@ -726,7 +726,7 @@ class DAQ_Scan(QObject):
         self.ui.scan_dock.addWidget(widgetsettings)
 
         self.dockarea.addDock(self.ui.scan_dock, 'left')
-        self.ui.scan_dock.float()
+        #self.ui.scan_dock.float()
 
         # %% init the 1D viewer
         self.ui.scan1D_graph_widget = QtWidgets.QSplitter(orientation=QtCore.Qt.Vertical)
